+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- Storage backend for ScyllaDB, a cassandra-like distributed key-value store.
+
 ### Changed
 - StoredDocument objects are now serialized with protobuf to increase speed and reduce storage
   consumption.